Add files via upload
stp1. Start
step 2. Input the binary number as a string
step 3. Initialize variables:
decimal to 0
binaryLength to the length of the input binary string
i to 0
step 4. While i is less than binaryLength, do the following:
Get the i-th character from the binary string and store it in digit
Calculate the power as binaryLength - 1 - i and store it in power
If digit is '1', do the following:
Calculate 2 to the power of power
Add the result to the decimal
Increment i by 1
step 5. Display "Decimal equivalent: " followed by the value of decimal
step 6. End